Split into multiple components
